% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gglmplot.R
\name{gglmplot}
\alias{gglmplot}
\title{gglmplot}
\usage{
gglmplot(model, type)
}
\arguments{
\item{model}{An object of type `lm` or `glm`}

\item{type}{Type of plot to be returned ("fitted_resid", "resid_hist", "qq", "scale_location")}
}
\value{
A a `ggplot2` object for visual diagnostic of model validity.
}
\description{
Provides visual model diagnostic plots with `ggplot2`.
}
\examples{
data(mtcars)
m1 <- lm(mpg ~ cyl + disp + hp, data = mtcars)
gglmplot(m1, type = "qq")
}
