% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_model_classes.R
\name{list_model_classes}
\alias{list_model_classes}
\title{list_model_classes}
\usage{
list_model_classes(...)
}
\arguments{
\item{...}{Currently ignored. For extendability.}
}
\value{
A character vector containing the possible model classes that 
\code{gglm} works with.
}
\description{
Returns the possible model classes that \code{gglm} works with.
}
\note{
Note that these are not always the exact name of the class that that 
can be used. This is due to how some methods are written in the packages
\code{gglm} imports. For example, the model class "merMod" refers to a 
variety of model outputs from \code{lme4}, even when the outputted class is
not called "merMod".
}
\examples{
list_model_classes()
}
