% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat_cooks_obs.R
\name{stat_cooks_obs}
\alias{stat_cooks_obs}
\title{stat_cooks_obs}
\usage{
stat_cooks_obs(mapping = ggplot2::aes(), ...)
}
\arguments{
\item{mapping}{Additional aesthetic mappings to add to the layer, e.g.
`aes(fill = some_var)`.
If `mapping` includes a required aesthetic (e.g. `x`, `y`), the supplied
value is ignored in favor of the default and a warning is issued.}

\item{...}{Additional arguments passed on to `geom_col()`.}
}
\value{
A `ggplot2` layer for plotting cook's distance by observation number.
}
\description{
`ggplot2` layer for plotting cook's distance by observation number.
}
\examples{
data(mtcars)
model <- lm(mpg ~ cyl + disp + hp, data = mtcars)
ggplot2::ggplot(data = model) + stat_cooks_obs()
}
