% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat_normal_qq.R
\name{stat_normal_qq}
\alias{stat_normal_qq}
\title{stat_normal_qq}
\usage{
stat_normal_qq(mapping = ggplot2::aes(), alpha = 0.5, ...)
}
\arguments{
\item{mapping}{Additional aesthetic mappings to add to the layer, e.g.
`aes(color = some_var)`.
If `mapping` includes a required aesthetic (e.g. `x`, `y`), the supplied
value is ignored in favor of the default and a warning is issued.}

\item{alpha}{Adjust transparency of points.}

\item{...}{Additional arguments passed on to `geom_point()`.}
}
\value{
A `ggplot2` layer for plotting a Normal Q-Q plot.
}
\description{
Normal QQ plot.
}
\examples{
data(mtcars)
model <- lm(mpg ~ cyl + disp + hp, data = mtcars)
ggplot2::ggplot(data = model) + stat_normal_qq()
}
