% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat_resid.R
\name{stat_resid}
\alias{stat_resid}
\title{stat_resid}
\usage{
stat_resid()
}
\value{
A a `ggplot2` object for visual diagnostic of model validity.
}
\description{
Visualize the residuals of a model versus their fitted values.
}
