% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gglm.R
\name{gglm}
\alias{gglm}
\title{gglm}
\usage{
gglm(data, mapping = ggplot2::aes(), theme = ggplot2::theme_gray(), ...)
}
\arguments{
\item{data}{A model object.}

\item{mapping}{Additional aesthetic mappings to add to each layer, e.g.
`aes(color = some_var)`.}

\item{theme}{The theme of the `ggplot`s to be produced.}

\item{...}{Additional arguments passed on to each `stat_*()` layer.}
}
\value{
A `ggplot2` object for visual diagnostic of model validity.
}
\description{
Provides four standard visual model diagnostic plots with `ggplot2`.
}
\examples{
data(mtcars)
m1 <- lm(mpg ~ cyl + disp + hp, data = mtcars)
gglm(m1)
}
