% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gglm.R
\name{gglm}
\alias{gglm}
\title{gglm}
\usage{
gglm(model)
}
\arguments{
\item{model}{An object of type `lm` or `glm`}
}
\value{
A a `ggplot2` object for visual diagnostic of model validity.
}
\description{
Provides four standard visual model diagnostic plots with `ggplot2`.
}
\examples{
library(patchwork)
data(mtcars)
m1 <- lm(mpg ~ cyl + disp + hp, data = mtcars)
gglm(m1)
}
