% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gglm.R
\name{gglm}
\alias{gglm}
\title{gglm}
\usage{
gglm(data, theme = ggplot2::theme_gray(), ...)
}
\arguments{
\item{data}{A model object of type `lm` or `glm`.}

\item{theme}{The theme of the `ggplot`s to be produced.}

\item{...}{Currently ignored. For extendability.}
}
\value{
A a `ggplot2` object for visual diagnostic of model validity.
}
\description{
Provides four standard visual model diagnostic plots with `ggplot2`.
}
\examples{
data(mtcars)
m1 <- lm(mpg ~ cyl + disp + hp, data = mtcars)
gglm(m1)
}
