% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gglm.R
\name{gglm}
\alias{gglm}
\title{Linear regression of two continuous variables with addition of R2 and p-value}
\usage{
gglm(
  data = iris,
  x = "Petal.Length",
  y = "Petal.Width",
  group = NULL,
  facet = TRUE,
  label.x = "left",
  label.y = "top",
  size = 14,
  show.se = TRUE
)
}
\arguments{
\item{data}{a data frame containing all the variables and the grouping information}

\item{x}{a variable in the x-axis}

\item{y}{a variable in the y-axis}

\item{group}{the grouping information}

\item{facet}{a boolean value to indicate facet}

\item{label.x}{the x position of label or can be "left" or "right" means the range of [0, 1]}

\item{label.y}{the y position og label or can be "bottom" or "top" means the range of [0, 1]}

\item{size}{the text size for ggtheme::theme_size function, default value is 14}

\item{show.se}{a boolean value to indicate whether shown the SE}
}
\value{
a ggplot2 plot
}
\description{
Linear regression of two continuous variables with addition of R2 and p-value
}
\examples{
gglm(group = "Species", label.y = c(1,1,75,2.6)) + ggsci::scale_color_aaas()
gglm(group = "Species", label.y = 0.99) + ggsci::scale_color_aaas() + gglm:::theme_pub()
gglm() + ggpubr::theme_pubr()
gglm() + egg::theme_article()
gglm() + ggprism::theme_prism()
gglm(group = "Species", label.y = 0.99) + ggsci::scale_color_aaas() + ggprism::theme_prism()
gglm(group = "Species", label.y = 0.99) + ggprism::theme_prism(base_size = 14) +
ggprism::scale_color_prism(palette = "floral")
}
\author{
Zhonghui Gai
}
