% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/launch.R
\name{launch}
\alias{launch}
\title{launch}
\usage{
launch(data = NULL, ...)
}
\arguments{
\item{data}{A `data.frame` to use in modeling.}

\item{...}{Currently ignored. For extendability.}
}
\value{
None. Launches a `shiny` applet for point-and-click based model
diagnostics.
}
\description{
Launches a R Shiny-based applet for linear model diagnostics.
}
\examples{
\dontrun{
launch()
}
}
