% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat_cooks_leverage.R
\name{stat_cooks_leverage}
\alias{stat_cooks_leverage}
\title{stat_cooks_leverage}
\usage{
stat_cooks_leverage(
  mapping = ggplot2::aes(),
  alpha = 0.5,
  method = "loess",
  linecolor = "steelblue",
  se = FALSE,
  ...
)
}
\arguments{
\item{mapping}{Additional aesthetic mappings to add to the layer, e.g.
\code{aes(color = some_var)}.
If \code{mapping} includes a required aesthetic (e.g. \code{x}, \code{y}), the
supplied value is ignored in favor of the default and a warning is issued.}

\item{alpha}{Adjust transparency of points.}

\item{method}{Method for fitting the line to the points.}

\item{linecolor}{Color of the line.}

\item{se}{Keep standard error bands around line?}

\item{...}{Additional arguments passed on to \code{geom_point()}.}
}
\value{
A \code{ggplot2} layer for plotting Cook's Distance vs. Leverage.
}
\description{
Cook's Distance vs. Leverage
}
\examples{
data(mtcars)
model <- lm(mpg ~ cyl + disp + hp, data = mtcars)
ggplot2::ggplot(data = model) + stat_cooks_leverage()
}
