% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat_cooks_obs.R
\name{stat_cooks_obs}
\alias{stat_cooks_obs}
\title{stat_cooks_obs}
\usage{
stat_cooks_obs(mapping = ggplot2::aes(), ...)
}
\arguments{
\item{mapping}{Additional aesthetic mappings to add to the layer, e.g.
\code{aes(fill = some_var)}.
If \code{mapping} includes a required aesthetic (e.g. \code{x}, \code{y}), the supplied
value is ignored in favor of the default and a warning is issued.}

\item{...}{Additional arguments passed on to \code{geom_col()}.}
}
\value{
A \code{ggplot2} layer for plotting cook's distance by observation number.
}
\description{
\code{ggplot2} layer for plotting cook's distance by observation number.
}
\examples{
data(mtcars)
model <- lm(mpg ~ cyl + disp + hp, data = mtcars)
ggplot2::ggplot(data = model) + stat_cooks_obs()
}
