% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat_cooks_obs.R
\name{stat_cooks_obs}
\alias{stat_cooks_obs}
\title{stat_cooks_obs}
\usage{
stat_cooks_obs(...)
}
\arguments{
\item{...}{Currently ignored. For extendability.}
}
\value{
A `ggplot2` layer for plotting cook's distance by observation number.
}
\description{
`ggplot2` layer for plotting cook's distance by observation number.
}
\examples{
data(mtcars)
model <- lm(mpg ~ cyl + disp + hp, data = mtcars)
ggplot2::ggplot(data = model) + stat_cooks_obs()
}
