% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat_fitted_resid.R
\name{stat_fitted_resid}
\alias{stat_fitted_resid}
\title{stat_fitted_resid}
\usage{
stat_fitted_resid(alpha = 0.5, ...)
}
\arguments{
\item{alpha}{Adjust transparency of points.}

\item{...}{Additional arguments to be passed to the `aes()` in `geom_point()`.}
}
\value{
A `ggplot2` layer for plotting a fitted vs. residual scatter plot.
}
\description{
`ggplot2` layer for plotting a fitted vs. residual scatter plot.
}
\examples{
data(mtcars)
model <- lm(mpg ~ cyl + disp + hp, data = mtcars)
ggplot2::ggplot(data = model) + stat_fitted_resid()
}
