% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat_fitted_resid.R
\name{stat_fitted_resid}
\alias{stat_fitted_resid}
\title{stat_fitted_resid}
\usage{
stat_fitted_resid(mapping = ggplot2::aes(), alpha = 0.5, ...)
}
\arguments{
\item{mapping}{Additional aesthetic mappings to add to the layer, e.g.
\code{aes(color = some_var)}.
If \code{mapping} includes a required aesthetic (e.g. \code{x}, \code{y}), the supplied
value is ignored in favor of the default and a warning is issued.}

\item{alpha}{Adjust transparency of points.}

\item{...}{Additional arguments passed on to \code{geom_point()}.}
}
\value{
A \code{ggplot2} layer for plotting a fitted vs. residual scatter plot.
}
\description{
\code{ggplot2} layer for plotting a fitted vs. residual scatter plot.
}
\examples{
data(mtcars)
model <- lm(mpg ~ cyl + disp + hp, data = mtcars)
ggplot2::ggplot(data = model) + stat_fitted_resid()
}
