% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat_fitted_resid.R
\name{stat_fitted_resid}
\alias{stat_fitted_resid}
\title{stat_fitted_resid}
\usage{
stat_fitted_resid()
}
\arguments{
\item{model}{An object of type `lm` or `glm`}
}
\value{
A a `ggplot2` object for visual diagnostic of model validity.
}
\description{
Provides four standard visual model diagnostic plots with `ggplot2`.
}
\examples{

}
