% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat_normal_qq.R
\name{stat_normal_qq}
\alias{stat_normal_qq}
\title{stat_normal_qq}
\usage{
stat_normal_qq(alpha = 0.5, ...)
}
\arguments{
\item{alpha}{Adjust transparency of points.}

\item{...}{Currently ignored. For extendability.}
}
\value{
A `ggplot2` layer for plotting a Normal Q-Q plot.
}
\description{
Normal QQ plot.
}
\examples{
data(mtcars)
model <- lm(mpg ~ cyl + disp + hp, data = mtcars)
ggplot2::ggplot(data = model) + stat_normal_qq()
}
