% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat_resid_hist.R
\name{stat_resid_hist}
\alias{stat_resid_hist}
\title{stat_resid_hist}
\usage{
stat_resid_hist(mapping = ggplot2::aes(), bins = 30, ...)
}
\arguments{
\item{mapping}{Additional aesthetic mappings to add to the layer, e.g.
\code{aes(fill = some_var)}.
If \code{mapping} includes a required aesthetic (e.g. \code{x}, \code{y}), the supplied
value is ignored in favor of the default and a warning is issued.}

\item{bins}{Adjust the number of bins.}

\item{...}{Additional arguments passed on to \code{geom_histogram()}.}
}
\value{
A \code{ggplot2} layer for plotting a histogram of residuals.
}
\description{
Visualize the distribution of the residuals of a model.
}
\examples{
data(mtcars)
model <- lm(mpg ~ cyl + disp + hp, data = mtcars)
ggplot2::ggplot(data = model) + stat_resid_hist()
}
