% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat_resid_hist.R
\name{stat_resid_hist}
\alias{stat_resid_hist}
\title{stat_resid_hist}
\usage{
stat_resid_hist(bins = 30)
}
\value{
A a `ggplot2` object for visual diagnostic of model validity.
}
\description{
Visualize the distribution of the residuals of a model.
}
