% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat_resid_hist.R
\name{stat_resid_hist}
\alias{stat_resid_hist}
\title{stat_resid_hist}
\usage{
stat_resid_hist(bins = 30)
}
\arguments{
\item{bins}{Default: 30. Adjust the number of bins.}
}
\value{
A `ggplot2` layer for plotting a histogram of residuals.
}
\description{
Visualize the distribution of the residuals of a model.
}
\examples{
data(mtcars)
model <- lm(mpg ~ cyl + disp + hp, data = mtcars)
model_tbl <- broom::augment(model)
ggplot2::ggplot(data = model_tbl) + stat_resid_hist()
}
