% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat_resid_leverage.R
\name{stat_resid_leverage}
\alias{stat_resid_leverage}
\title{stat_resid_leverage}
\usage{
stat_resid_leverage(
  mapping = ggplot2::aes(),
  alpha = 0.5,
  method = "loess",
  se = FALSE,
  linecolor = "steelblue",
  ...
)
}
\arguments{
\item{mapping}{Additional aesthetic mappings to add to the layer, e.g.
\code{aes(color = some_var)}.
If \code{mapping} includes a required aesthetic (e.g. \code{x}, \code{y}), the supplied
value is ignored in favor of the default and a warning is issued.}

\item{alpha}{Adjust transparency of points.}

\item{method}{Method for fitting the line to the points.}

\item{se}{Keep standard error bands around line?}

\item{linecolor}{Color of the line.}

\item{...}{Additional arguments passed on to \code{geom_point()}.}
}
\value{
A \code{ggplot2} layer for plotting a fitted vs. residual scatter plot.
}
\description{
Residual vs. leverage plot.
}
\examples{
data(mtcars)
model <- lm(mpg ~ cyl + disp + hp, data = mtcars)
ggplot2::ggplot(data = model) + stat_resid_leverage()
}
