% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat_scale_location.R
\name{stat_scale_location}
\alias{stat_scale_location}
\title{stat_scale_location}
\usage{
stat_scale_location(
  mapping = ggplot2::aes(),
  alpha = 0.5,
  na.rm = TRUE,
  se = FALSE,
  method = "loess",
  linecolor = "steelblue",
  ...
)
}
\arguments{
\item{mapping}{Additional aesthetic mappings to add to the layer, e.g.
`aes(color = some_var)`.
If `mapping` includes a required aesthetic (e.g. `x`, `y`), the supplied
value is ignored in favor of the default and a warning is issued.}

\item{alpha}{Adjust the transparency of points.}

\item{na.rm}{Remove points with value NA?}

\item{se}{Keep standard error bands around line?}

\item{method}{Method for fitting the line to the points.}

\item{linecolor}{Color of the line.}

\item{...}{Additional arguments passed on to `geom_point()`.}
}
\value{
A `ggplot2` layer for plotting the scale location diagnostic plot.
}
\description{
Scale location diagnostic plot.
}
\examples{
data(mtcars)
model <- lm(mpg ~ cyl + disp + hp, data = mtcars)
ggplot2::ggplot(data = model) + stat_scale_location()
}
