% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat_scale_location.R
\name{stat_scale_location}
\alias{stat_scale_location}
\title{stat_scale_location}
\usage{
stat_scale_location()
}
\arguments{
\item{model}{An object of type `lm` or `glm`}
}
\value{
A a `ggplot2` object for visual diagnostic of model validity.
}
\description{
Provides four standard visual model diagnostic plots with `ggplot2`.
}
\examples{

}
